Moved CaptureTransmitter
